fix: roll back CodeAnalysis version
CodeAnalysis 4.11.0 is the latest version available for .NET 8, and the
recent bump to 4.14.0 causes warnings about compiler-version mismatches
in consuming projects using .NET 8.

This change rolls back the Analyzers project's CodeAnalysis references
to 4.11.0, but marks the Workspaces reference with PrivateAssets="all"
to avoid transitive references causing slow builds in consuming
projects.

It also rolls back the BenchmarkDotNet version in
Analyzers.PerformanceTests for compatibility with CodeAnalysis 4.11.0
and re-marks BenchmarkDotNet as ignored in renovate.